better emlog api
I made private the functions which perform the actuall printing of logging messages (i.e. print, printf and println). I also improved the NewEmersyxLogger method, by making it the problem of the caller to supply an io.Writer instance where logs should be written to. Additionally, I added the feature of printing the file name and line number to logs, if this information is available (by using runtime.Caller).